Explicitly cast negative RepeatCount::Max value to unsigned int
The scanner and parser we use internally make use of a tokenizer which
retrieves a maximum repeat count for repeating tokens. In order to
signal "unlimited" we used a negative value of an unsigned int, which
wraps around to the maximum value representable by this data type.
However, clang complains about this happening implicitly and so with
this change we make the cast explicit.